Comments, some new functionality
Commented every page for submission to the client. Also added two small
bits of functionality: the login page now includes a hint that your
username is the first part of your Algonquin email address; and the
registration email now tells you your username. Also added code to allow
unauthorized page accesses to be logged in a database table, but left
this commented out -- this is just an option that the client might want
to use.